+++ lunabot_behavior/excavate.py
@@ -10,7 +10,7 @@ from pid_controller import VelocityPIDController
 from clamp_output import clamp_output  
 import ascent
 
-from std_msgs.msg import Int8
+from std_msgs.msg import Int8, Int32
 
 
 class Excavate:
@@ -34,7 +34,7 @@ class Excavate:
         """
 
         if excavation_publisher is None:
-            self.excavation_publisher: rospy.Publisher = rospy.Publisher("/excavate", Int8, queue_size=1, latch=True)
+            self.excavation_publisher: rospy.Publisher = rospy.Publisher("/excavate", Int32, queue_size=1, latch=True)
             rospy.init_node("excavation_node")
         else:
             self.excavation_publisher: rospy.Publisher = excavation_publisher
@@ -58,6 +58,7 @@ class Excavate:
         # 90 percent of max speed
         TARGET_EXCAVATION_VELOCITY = 8 * 0.9
 
+        # TODO RJN - update this whole script to use RPM
         self.excavation_pid_controller = VelocityPIDController(
             TARGET_EXCAVATION_VELOCITY, 1, 0, 0, 127
         )  # TODO find values
@@ -111,7 +112,7 @@ class Excavate:
             time.sleep(3)
             return True
 
-        excavation_message = Int8()
+        excavation_message = Int32()
         lin_act_message = Int8()
 
         excavation_ang = (
@@ -190,7 +191,7 @@ class Excavate:
             time.sleep(3)
             return True
 
-        excavation_message = Int8()
+        excavation_message = Int32()
         cmd_vel_message = Twist()
 
         excavation_ang = self.robot_sensors.exc_ang 
@@ -276,7 +277,7 @@ class Excavate:
         # 90% of max speed, backwards
         EXCAVATION_SPEED = int(-127 * 0.9)
 
-        excavation_message = Int8()
+        excavation_message = Int32()
         excavation_message.data = EXCAVATION_SPEED
 
         self.excavation_publisher.publish(excavation_message)
